requirements.txt
langchain_community
langchainhub
langchain_chroma
langchain
langgraph
python-dotenv
langchain-groq
langchain_huggingface
beautifulsoup4